mcr-settlement-config: don't use u128
Work around TOML serialization issues for u128,
by changing the type for gas_limit to u64.

+++ protocol-units/settlement/mcr/client/src/eth_client.rs
@@ -102,7 +102,7 @@ impl
 			ws_url,
 			signer_address,
 			contract_address,
-			config.gas_limit,
+			config.gas_limit as u128,
 			config.tx_send_retries,
 		)
 		.await
